+++ src/plume/PI_ReleaseType_continuous.hpp
@@ -43,8 +43,8 @@ class PI_ReleaseType_continuous : public PI_ReleaseType
 {
 private:
   // note that this also inherits data members:
-  // ParticleReleaseType m_rType, int m_particlePerTimestep, double m_releaseStartTime,
-  //  double m_releaseEndTime, and int m_numPar from ReleaseType.
+  // ParticleReleaseType m_rType, int m_particlePerTimestep, float m_releaseStartTime,
+  //  float m_releaseEndTime, and int m_numPar from ReleaseType.
   // guidelines for how to set these variables within an inherited ReleaseType are given in ReleaseType.hpp.
 
 
@@ -61,11 +61,11 @@ public:
   void parseValues() override
   {
     parsePrimitive<int>(true, m_particlePerTimestep, "particlePerTimestep");
-    parsePrimitive<double>(false, m_massPerSec, "massPerSec");
+    parsePrimitive<float>(false, m_massPerSec, "massPerSec");
   }
 
 
-  void calcReleaseInfo(const double &timestep, const double &simDur) override
+  void calcReleaseInfo(const float &timestep, const float &simDur) override
   {
     // set the overall releaseType variables from the variables found in this class
     m_releaseStartTime = 0;
