+++ src/plume/PI_SourceGeometry_FullDomain.cpp
@@ -37,7 +37,12 @@
 
 #include "PI_SourceGeometry_FullDomain.hpp"
 
-void PI_SourceGeometry_FullDomain::checkPosInfo(const double &domainXstart, const double &domainXend, const double &domainYstart, const double &domainYend, const double &domainZstart, const double &domainZend)
+void PI_SourceGeometry_FullDomain::checkPosInfo(const float &domainXstart,
+                                                const float &domainXend,
+                                                const float &domainYstart,
+                                                const float &domainYend,
+                                                const float &domainZstart,
+                                                const float &domainZend)
 {
 
   // notice that setting the variables as I am doing right now is not the standard way of doing this function
